few extra lines, added, plus comments

+++ src/LeviThompsonLab02/Measurement.java
@@ -14,15 +14,15 @@ public class Measurement {
     double value;
     String unit;
     //method to assign new values to attribute when object is created
-    public Measurement(double value, String unit) {
+    public Measurement(int value, String unit) {
         this.value = value;
         this.unit = unit;
-                
+        
     }
 
     @Override
     public String toString() {
-        return value + unit;
+        return (int)value + unit;
     }
 
 }
